Update 1.1 tests (#63)
* Update tests for saml merge

* Remove authnz_test from old location
