+++ setup.cfg
@@ -14,7 +14,7 @@ with-coverage = 1
 cover-package = confidant
 cover-xml = 1
 cover-xml-file = build/coverage.xml
-cover-min-percentage = 40
+cover-min-percentage = 55
 
 [flake8]
 # The jenkins violations plugin can read the pylint format.
